Add case weight support for `mcc()`

+++ src/mcc-multiclass.c
@@ -11,7 +11,7 @@
 static inline R_len_t compute_flat_index(int row, int col, R_len_t n_row);
 
 SEXP yardstick_mcc_multiclass_impl(SEXP C) {
-  const int* p_C = INTEGER(C);
+  const double* p_C = REAL(C);
 
   // `C` is square so `n_row == n_col`
   SEXP dim = PROTECT(Rf_getAttrib(C, R_DimSymbol));
@@ -36,7 +36,7 @@ SEXP yardstick_mcc_multiclass_impl(SEXP C) {
   for(int k = 0; k < N; ++k) {
     const R_len_t loc_kk = compute_flat_index(k, k, N);
 
-    const double C_kk = (double) p_C[loc_kk];
+    const double C_kk = p_C[loc_kk];
 
     // reset to 0 for each k iteration
     C_lk_sum = 0;
@@ -50,15 +50,15 @@ SEXP yardstick_mcc_multiclass_impl(SEXP C) {
       const R_len_t loc_lk = compute_flat_index(l, k, N);
       const R_len_t loc_kl = compute_flat_index(k, l, N);
 
-      const double C_lk = (double) p_C[loc_lk];
-      const double C_kl = (double) p_C[loc_kl];
+      const double C_lk = p_C[loc_lk];
+      const double C_kl = p_C[loc_kl];
 
       for(int m = 0; m < N; ++m) {
         const R_len_t loc_ml = compute_flat_index(m, l, N);
         const R_len_t loc_km = compute_flat_index(k, m, N);
 
-        const double C_ml = (double) p_C[loc_ml];
-        const double C_km = (double) p_C[loc_km];
+        const double C_ml = p_C[loc_ml];
+        const double C_km = p_C[loc_km];
 
         // Actual increment of numerator sum
         numer += C_kk * C_ml - C_lk * C_km;
@@ -81,8 +81,8 @@ SEXP yardstick_mcc_multiclass_impl(SEXP C) {
         const R_len_t loc_gf = compute_flat_index(g, f, N);
         const R_len_t loc_fg = compute_flat_index(f, g, N);
 
-        C_gf_sum += (double) p_C[loc_gf];
-        C_fg_sum += (double) p_C[loc_fg];
+        C_gf_sum += p_C[loc_gf];
+        C_fg_sum += p_C[loc_fg];
       }
     }
